+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -239,9 +239,6 @@ public enum FeatureFlag: String, CaseIterable {
     /// Uses the PlaylistMetadataLoader cache before running the query (the query will update when it's done)
     case playlistDataCacheBeforeQuery
 
-    /// Avoid returning cached episode early and use policy instead
-    case episodesInfoCacheReloadPolicy
-
     /// Ignores play remote commands when other audio is playing
     case ignorePlayWithOtherAudio
 
@@ -457,8 +454,6 @@ public enum FeatureFlag: String, CaseIterable {
             true
         case .playlistDataCacheBeforeQuery:
             true
-        case .episodesInfoCacheReloadPolicy:
-			true
         case .ignorePlayWithOtherAudio:
             true
         case .streamAndDownloadReadFromMemoryBuffer: